requirements.txt
discord.py
pyombi
